requirements.txt
# requirements.txt
streamlit==1.35.0
langchain==0.2.5
langchain-google-genai==0.0.17
python-dotenv==1.0.1
requests==2.32.3
google-api-python-client==2.130.0
chromadb==0.5.3
unstructured==0.14.4
beautifulsoup4==4.12.3
pytest==8.2.2
pytest-mock==3.12.0